# Refreshing VM performance by separating verification and tweaking buffer access.