--- a/compiler/src/iree/compiler/Dialect/VM/Conversion/UtilToVM/ConvertBufferOps.cpp
+++ b/compiler/src/iree/compiler/Dialect/VM/Conversion/UtilToVM/ConvertBufferOps.cpp
@@ -154,6 +154,14 @@ struct BufferCompareOpConversion
   }
 };
 
+static Value unscaleOffset(Location loc, Value offset, int64_t scale,
+                           OpBuilder &builder) {
+  if (scale == 1) return offset;
+  return builder.createOrFold<IREE::VM::DivI64SOp>(
+      loc, offset.getType(), offset,
+      builder.create<IREE::VM::ConstI64Op>(loc, scale));
+}
+
 struct BufferFillOpConversion
     : public OpConversionPattern<IREE::Util::BufferFillOp> {
   using OpConversionPattern::OpConversionPattern;
@@ -168,6 +176,11 @@ struct BufferFillOpConversion
     }
     auto byteOffset = castToI64(adaptor.getTargetOffset(), rewriter);
     auto byteLength = castToI64(adaptor.getLength(), rewriter);
+    int64_t elementSize = IREE::Util::getRoundedElementByteWidth(oldType);
+    auto elementOffset =
+        unscaleOffset(fillOp.getLoc(), byteOffset, elementSize, rewriter);
+    auto elementLength =
+        unscaleOffset(fillOp.getLoc(), byteLength, elementSize, rewriter);
     auto pattern = adaptor.getPattern();
     if (auto integerType = oldType.dyn_cast<IntegerType>()) {
       if (integerType.isInteger(1) || integerType.isInteger(8)) {
@@ -175,23 +188,23 @@ struct BufferFillOpConversion
             fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
       } else if (integerType.isInteger(16)) {
         rewriter.replaceOpWithNewOp<IREE::VM::BufferFillI16Op>(
-            fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
+            fillOp, adaptor.getTarget(), elementOffset, elementLength, pattern);
       } else if (integerType.isInteger(32)) {
         rewriter.replaceOpWithNewOp<IREE::VM::BufferFillI32Op>(
-            fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
+            fillOp, adaptor.getTarget(), elementOffset, elementLength, pattern);
       } else if (integerType.isInteger(64)) {
         rewriter.replaceOpWithNewOp<IREE::VM::BufferFillI64Op>(
-            fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
+            fillOp, adaptor.getTarget(), elementOffset, elementLength, pattern);
       } else {
         return rewriter.notifyMatchFailure(
             fillOp, "invalid integer buffer element type");
       }
     } else if (oldType.isF32()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferFillF32Op>(
-          fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
+          fillOp, adaptor.getTarget(), elementOffset, elementLength, pattern);
     } else if (oldType.isF64()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferFillF64Op>(
-          fillOp, adaptor.getTarget(), byteOffset, byteLength, pattern);
+          fillOp, adaptor.getTarget(), elementOffset, elementLength, pattern);
     } else {
       return rewriter.notifyMatchFailure(fillOp,
                                          "invalid float buffer element type");
@@ -199,14 +212,6 @@ struct BufferFillOpConversion
     return success();
   }
 };
-
-static Value unscaleOffset(Location loc, Value offset, int64_t scale,
-                           OpBuilder &builder) {
-  if (scale == 1) return offset;
-  return builder.createOrFold<IREE::VM::DivI64SOp>(
-      loc, offset.getType(), offset,
-      builder.create<IREE::VM::ConstI64Op>(loc, scale));
-}
 
 struct BufferLoadOpConversion
     : public OpConversionPattern<IREE::Util::BufferLoadOp> {
@@ -220,6 +225,9 @@ struct BufferLoadOpConversion
       oldType = newType;
     }
     auto byteOffset = castToI64(adaptor.getSourceOffset(), rewriter);
+    int64_t elementSize = IREE::Util::getRoundedElementByteWidth(oldType);
+    auto elementOffset =
+        unscaleOffset(loadOp.getLoc(), byteOffset, elementSize, rewriter);
     if (auto integerType = oldType.dyn_cast<IntegerType>()) {
       if (integerType.isInteger(1) || integerType.isInteger(8)) {
         if (integerType.isSigned() || integerType.isSignless()) {
@@ -232,33 +240,27 @@ struct BufferLoadOpConversion
       } else if (integerType.isInteger(16)) {
         if (integerType.isSigned() || integerType.isSignless()) {
           rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadI16SOp>(
-              loadOp, newType, adaptor.getSource(),
-              unscaleOffset(loadOp.getLoc(), byteOffset, 2, rewriter));
+              loadOp, newType, adaptor.getSource(), elementOffset);
         } else {
           rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadI16UOp>(
-              loadOp, newType, adaptor.getSource(),
-              unscaleOffset(loadOp.getLoc(), byteOffset, 2, rewriter));
+              loadOp, newType, adaptor.getSource(), elementOffset);
         }
       } else if (integerType.isInteger(32)) {
         rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadI32Op>(
-            loadOp, newType, adaptor.getSource(),
-            unscaleOffset(loadOp.getLoc(), byteOffset, 4, rewriter));
+            loadOp, newType, adaptor.getSource(), elementOffset);
       } else if (integerType.isInteger(64)) {
         rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadI64Op>(
-            loadOp, newType, adaptor.getSource(),
-            unscaleOffset(loadOp.getLoc(), byteOffset, 8, rewriter));
+            loadOp, newType, adaptor.getSource(), elementOffset);
       } else {
         return rewriter.notifyMatchFailure(
             loadOp, "invalid integer buffer element type");
       }
     } else if (oldType.isF32()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadF32Op>(
-          loadOp, newType, adaptor.getSource(),
-          unscaleOffset(loadOp.getLoc(), byteOffset, 4, rewriter));
+          loadOp, newType, adaptor.getSource(), elementOffset);
     } else if (oldType.isF64()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferLoadF64Op>(
-          loadOp, newType, adaptor.getSource(),
-          unscaleOffset(loadOp.getLoc(), byteOffset, 8, rewriter));
+          loadOp, newType, adaptor.getSource(), elementOffset);
     } else {
       return rewriter.notifyMatchFailure(loadOp,
                                          "invalid float buffer element type");
@@ -279,34 +281,27 @@ struct BufferStoreOpConversion
       oldType = newType;
     }
     auto byteOffset = castToI64(adaptor.getTargetOffset(), rewriter);
+    int64_t elementSize = IREE::Util::getRoundedElementByteWidth(oldType);
+    auto elementOffset =
+        unscaleOffset(storeOp.getLoc(), byteOffset, elementSize, rewriter);
     if (oldType.isInteger(1) || oldType.isInteger(8)) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreI8Op>(
           storeOp, adaptor.getTarget(), byteOffset, adaptor.getSource());
     } else if (oldType.isInteger(16)) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreI16Op>(
-          storeOp, adaptor.getTarget(),
-          unscaleOffset(storeOp.getLoc(), byteOffset, 2, rewriter),
-          adaptor.getSource());
+          storeOp, adaptor.getTarget(), elementOffset, adaptor.getSource());
     } else if (oldType.isInteger(32)) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreI32Op>(
-          storeOp, adaptor.getTarget(),
-          unscaleOffset(storeOp.getLoc(), byteOffset, 4, rewriter),
-          adaptor.getSource());
+          storeOp, adaptor.getTarget(), elementOffset, adaptor.getSource());
     } else if (oldType.isInteger(64)) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreI64Op>(
-          storeOp, adaptor.getTarget(),
-          unscaleOffset(storeOp.getLoc(), byteOffset, 8, rewriter),
-          adaptor.getSource());
+          storeOp, adaptor.getTarget(), elementOffset, adaptor.getSource());
     } else if (oldType.isF32()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreF32Op>(
-          storeOp, adaptor.getTarget(),
-          unscaleOffset(storeOp.getLoc(), byteOffset, 4, rewriter),
-          adaptor.getSource());
+          storeOp, adaptor.getTarget(), elementOffset, adaptor.getSource());
     } else if (oldType.isF64()) {
       rewriter.replaceOpWithNewOp<IREE::VM::BufferStoreF64Op>(
-          storeOp, adaptor.getTarget(),
-          unscaleOffset(storeOp.getLoc(), byteOffset, 8, rewriter),
-          adaptor.getSource());
+          storeOp, adaptor.getTarget(), elementOffset, adaptor.getSource());
     } else {
       return rewriter.notifyMatchFailure(storeOp,
                                          "invalid buffer element type");