--- a/compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/BytecodeEncoder.h
+++ b/compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/BytecodeEncoder.h
@@ -18,8 +18,13 @@ namespace IREE {
 namespace VM {
 
 struct EncodedBytecodeFunction {
-  // Encoded bytecode data for the function body.
+  // Encoded bytecode data for the function body including padding.
   std::vector<uint8_t> bytecodeData;
+  // Precise size of the bytecode.
+  size_t bytecodeLength = 0;
+
+  // Total number of blocks including the entry block.
+  uint16_t blockCount = 0;
 
   // Total i32 register slots required for execution.
   // Note that larger types also use these slots (i64=2xi32).
@@ -32,7 +37,7 @@ struct EncodedBytecodeFunction {
 class BytecodeEncoder : public VMFuncEncoder {
  public:
   // Matches IREE_VM_BYTECODE_VERSION_MAJOR.
-  static constexpr uint32_t kVersionMajor = 13;
+  static constexpr uint32_t kVersionMajor = 14;
   // Matches IREE_VM_BYTECODE_VERSION_MINOR.
   static constexpr uint32_t kVersionMinor = 0;
   static constexpr uint32_t kVersion = (kVersionMajor << 16) | kVersionMinor;