--- a/compiler/src/iree/compiler/Dialect/VM/Transforms/Passes.cpp
+++ b/compiler/src/iree/compiler/Dialect/VM/Transforms/Passes.cpp
@@ -69,7 +69,7 @@ void buildVMTransformPassPipeline(OpPassManager &passManager,
   FunctionLikeNest(passManager)
       .addPass(mlir::createSCFForLoopCanonicalizationPass);
 
-  // This pass is sketchy as it can pessimizes tight loops due to affine
+  // This pass is sketchy as it can pessimize tight loops due to affine
   // treating all indices as signed and the unsigned conversion pass not being
   // able to handle that. The scf.for canonicalization does a decent job of
   // removing trivial loops above and this catches the rest. It inserts nasty