… build_tools/cmake/iree_copts.cmake
… compiler/src/iree/compiler/Dialect/VM/Conversion/UtilToVM/test/buffer_ops.mlir
… compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/BytecodeModuleTarget.cpp
… compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/CMakeLists.txt
… compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/test/module_encoding_smoke.mlir
… runtime/src/iree/base/internal/debugging.h
… runtime/src/iree/base/target_platform.h
… runtime/src/iree/modules/hal/types.c
… runtime/src/iree/schemas/bytecode_module_def.fbs
… runtime/src/iree/tooling/trace_replay.h
… runtime/src/iree/vm/buffer.c
… runtime/src/iree/vm/buffer.h
… runtime/src/iree/vm/bytecode/BUILD
… runtime/src/iree/vm/bytecode/CMakeLists.txt
… runtime/src/iree/vm/bytecode/archive.c
… runtime/src/iree/vm/bytecode/archive.h
… runtime/src/iree/vm/bytecode/disassembler.c
… runtime/src/iree/vm/bytecode/dispatch.c
… runtime/src/iree/vm/bytecode/dispatch_util.h
… runtime/src/iree/vm/bytecode/generated/op_table.h
… runtime/src/iree/vm/bytecode/module.c
… runtime/src/iree/vm/bytecode/module.h
… runtime/src/iree/vm/bytecode/module_benchmark.mlir
… runtime/src/iree/vm/bytecode/module_impl.h
… runtime/src/iree/vm/bytecode/utils/BUILD
… runtime/src/iree/vm/bytecode/utils/CMakeLists.txt
… runtime/src/iree/vm/bytecode/utils/block_list.c
… runtime/src/iree/vm/bytecode/utils/block_list.h
… runtime/src/iree/vm/bytecode/utils/block_list_test.cc
… runtime/src/iree/vm/bytecode/utils/features.c